// fake-node/service/src/lib/common/types/brand-name.ts
export enum BrandName {
  AllBet = 'AllBet',
  AsiaGaming = 'AsiaGaming',
  DreamGaming = 'DreamGaming',
  Fake = 'Fake',
  Evolution = 'Evolution',
  //Evolution = 'Evolution',
  OrientalGaming = 'OrientalGaming',
  Wanmei = 'Wanmei',
  Sexy = 'Sexy',
  Bota = 'Bota',
  BotaKhan = 'BotaKhan',
  MegaHoldem = 'MegaHoldem',

  Dowinn = 'Dowinn',

  LiveBar = 'LiveBar',
  Univerge = 'Univerge',
  NetEnt = 'NetEnt',
  BigTimeGaming = 'BigTimeGaming',
  RedTiger = 'RedTiger',
  PragmaticPlay = 'PragmaticPlay',
  Habanero = 'Habanero',
  Playson = 'Playson',
  Booongo = 'Booongo',
  XINGaming = 'XINGaming',
  Wazdan = 'Wazdan',
  RelaxGaming = 'RelaxGaming',
  GameArt = 'GameArt',
  BlueprintGaming = 'BlueprintGaming',
  ThunderKick = 'ThunderKick',
  NoLimitCity = 'NoLimitCity',
  Mobilots = 'Mobilots',
  PlayPearls = 'PlayPearls',
  _1x2Gaming = '1x2Gaming',
  ElkStudios = 'ElkStudios',
  CX = 'CX',
  CG = 'CG',
  PlaynGO = `PlaynGO`,
  DragoonSoft = `DragoonSoft`,
  WorldEntertainment = 'WorldEntertainment',
  MotivationGaming = 'MotivationGaming',
  Yaxing = 'Yaxing',
  CQ9 = 'CQ9',

  MicroGaming = 'MicroGaming',
  SpiniX = 'SpiniX',
  Yggdrasil = 'Yggdrasil',
  NagaGames = 'NagaGames',
  TopTrendGaming = 'TopTrendGaming',
  LGDGaming = 'LGDGaming',

  QueenMaker = 'QueenMaker',
  OneGame = 'OneGame',
  KingMaker = 'KingMaker',
  BigGaming = 'BigGaming',
  Endorphina = 'Endorphina',
  PGSoft = 'PGSoft',
  Betixon = 'Betixon',
  PlayStar = 'PlayStar',
  SmartSoft = 'SmartSoft',
  GeminiGaming = 'GeminiGaming',
  PopOk = 'PopOk',
  VivoGaming = 'VivoGaming',
  BTI = 'BTISport',

  Live22 = 'Live22',

  Test = 'Test',
}